GS: Support fractional upscale

+++ pcsx2/GS/Renderers/HW/GSRendererHW.h
@@ -177,7 +177,7 @@ public:
 
 	void SetGameCRC(u32 crc, int options) override;
 	bool CanUpscale() override;
-	int GetUpscaleMultiplier() override;
+	float GetUpscaleMultiplier() override;
 	void Lines2Sprites();
 	bool VerifyIndices();
 	template <GSHWDrawConfig::VSExpand Expand> void ExpandIndices();
